+++ src/index.ts
@@ -84,6 +84,9 @@ export * as nip26 from './nips/nip-26.js';
 export * as nip44 from './nips/nip-44.js';
 export * as nip46 from './nips/nip-46.js';
 export * as nip49 from './nips/nip-49.js';
+export * as nip59 from './nips/nip-59.js';
+export * as nip17 from './nips/nip-17.js';
+export * as nip98 from './nips/nip-98.js';
 
 // Utils
 export {